+++ bee-storage/bee-storage/src/access/mod.rs
@@ -1,6 +1,9 @@
 // Copyright 2020-2021 IOTA Stiftung
 // SPDX-License-Identifier: Apache-2.0
 
+//! Access module which form the access layer of the backend which holds the contract of unified database access
+//! operations across all the backends and bee types.
+
 /// Holds the contract for batch access operation.
 mod batch;
 /// Holds the contract for delete access operation.