+++ bee-storage/bee-storage/src/backend.rs
@@ -1,6 +1,11 @@
 // Copyright 2020-2021 IOTA Stiftung
 // SPDX-License-Identifier: Apache-2.0
 
+//! Backend module which form the backend layer of the backend which holds the contract of starting and shutting down
+//! the backend.
+
+use crate::health::StorageHealth;
+
 use async_trait::async_trait;
 use serde::de::DeserializeOwned;
 
@@ -21,6 +26,12 @@ pub trait StorageBackend: Send + Sized + Sync + 'static {
     /// Shutdowns the backend.
     async fn shutdown(self) -> Result<(), Self::Error>;
 
-    /// Size of the database in bytes. Not all backends may be able to provide it, hence the option.
+    /// Returns the size of the database in bytes. Not all backends may be able to provide it, hence the option.
     async fn size(&self) -> Result<Option<usize>, Self::Error>;
+
+    /// Returns the health status of the database.
+    async fn get_health(&self) -> Result<Option<StorageHealth>, Self::Error>;
+
+    /// Sets the health status of the database.
+    async fn set_health(&self, health: StorageHealth) -> Result<(), Self::Error>;
 }